+++ backend/src/api/main_jh.py
@@ -4,6 +4,8 @@ from pydantic import BaseModel
 
 from src.core.spark import conn, connection_properties
 from src.core.settings import settings
+from src.core.queries import Jh_DRUNK_INFO_SQL, Jh_GET_STATION_SQL, search_complex_sql1, search_complex_sql2, search_complex_sql3,search_complex_sql4
+
 
 router = APIRouter(tags=["jihwan"])
 
@@ -39,19 +41,7 @@ def drunk_info(year:str):
   mssage = None
   try:    
     if getMetros(spark, year):
-      sql1 = """
-          SELECT
-            `역명`,
-            Floor(AVG(CAST(`20~21` AS INT) + CAST(`21~22` AS INT) + CAST(`22~23` AS INT)), 0) AS `night_avg`
-          FROM jhYearTable
-          WHERE (`주말여부` = true
-            OR DAYOFWEEK(CAST(`날짜` AS DATE)) = 6)
-            AND `구분` = '승차'
-          GROUP BY `역명`
-          ORDER BY night_avg DESC
-          LIMIT 10
-          """
-      fIdDf = spark.sql(sql1)
+      fIdDf = spark.sql(Jh_DRUNK_INFO_SQL)
       result_data = fIdDf.toPandas().to_dict(orient="records")
 
     # return {"status": True}
@@ -70,13 +60,7 @@ def get_station(year:str):
   mssage = None
   try:
     if getMetros(spark, year):
-      sql1 = """
-          SELECT
-            `역명`
-          FROM jhYearTable
-          GROUP BY `역명`
-          """
-      fIdDf = spark.sql(sql1)
+      fIdDf = spark.sql(Jh_GET_STATION_SQL)
       result_data = fIdDf.toPandas().to_dict(orient="records")
       station_list = []
       for data in result_data:
@@ -116,38 +100,15 @@ def search_complex(model: SearchQuery):
         for time in range(model.time, model.time+3) :
           if len(str(time)) == 1: timeList.append("0"+str(time))
           else : timeList.append(str(time))
-        
-        sql1 = f"""
-          SELECT `역명`, ROUND(AVG(CAST(`{timeList[0]}~{timeList[1]}` AS INT) + CAST(`{timeList[1]}~{timeList[2]}` AS INT)),0) AS `지정 평균 승객` 
-          FROM jhYearTable
-          WHERE `역명` IN ({stations_formatted}) 
-          GROUP BY `역명`;
-            """
-        fIdDf1 = spark.sql(sql1)
+        fIdDf1 = spark.sql(search_complex_sql1(timeList,stations_formatted))
         result_data1 = fIdDf1.toPandas().to_dict(orient="records")
-
-        sql2 = f'''
-        SELECT  ROUND(AVG(CAST(`{timeList[0]}~{timeList[1]}` AS INT) + CAST(`{timeList[1]}~{timeList[2]}` AS INT)),0) AS `전체 평균 승객` 
-        FROM jhYearTable
-        '''
-        fIdDf2 = spark.sql(sql2)
+        fIdDf2 = spark.sql(search_complex_sql2(timeList))
         result_data2 = fIdDf2.toPandas().to_dict(orient="records")
 
       else : 
-        sql1 = f"""
-          SELECT `역명`, ROUND(AVG(CAST(`{time}~{time+1}` AS INT)),0) AS `지정 평균 승객` 
-          FROM jhYearTable
-          WHERE `역명` IN ({stations_formatted}) 
-          GROUP BY `역명`;
-            """
-        fIdDf1 = spark.sql(sql1)
+        fIdDf1 = spark.sql(search_complex_sql3(model.time,stations_formatted))
         result_data1 = fIdDf1.toPandas().to_dict(orient="records")
-
-        sql2 = f'''
-        SELECT  ROUND(AVG(CAST(`{time}~{time+1}` AS INT)),0) AS `전체 평균 승객` 
-        FROM jhYearTable
-        '''
-        fIdDf2 = spark.sql(sql2)
+        fIdDf2 = spark.sql(search_complex_sql4(model.time))
         result_data2 = fIdDf2.toPandas().to_dict(orient="records")
 
   except Exception as e:
