+++ frontend/src/utils/network.js
@@ -1,7 +1,7 @@
 import axios from "axios"
 
 export const api = axios.create({
-  baseURL: import.meta.env.VITE_APP_FASTAPI_URL || "http://aiedu.tplinkdns.com:6011",
+  baseURL: import.meta.env.VITE_APP_FASTAPI_URL,
   // withCredentials: true,
   headers: {
     "Content-Type": "application/json"
@@ -9,7 +9,7 @@ export const api = axios.create({
 })
 
 export const n8n_api = axios.create({
-  baseURL: import.meta.env.VITE_APP_FASTAPI_URL || "http://aiedu.tplinkdns.com:7242",
+  baseURL: import.meta.env.VITE_APP_N8N_URL,
   // withCredentials: true,
   headers: {
     "Content-Type": "application/json"
